backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
distro==1.9.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.11.0
moviepy==2.2.1
numpy==2.3.3
openai==1.109.1
pillow==11.3.0
proglog==0.1.12
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytube==15.0.0
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
